pyproject.toml
[tool.poetry]
name = "dismissal-prediction-service"
version = "0.1.0"
description = "A service for predicting the imminent dismissal of each employee depending on events when working with corporate email"
authors = ["Parutin Maxim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.23"
sqlalchemy-utils = "^0.41.1"
pandas = "^2.1.4"
numpy = "^1.26.2"
opendatasets = "^0.1.22"
hydra-core = "^1.3.2"
googletrans = "3.1.0a0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
